prod/backend/requirements.txt
# FastAPI Backend Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.23
alembic>=1.13.0
psycopg2-binary>=2.9.9  # PostgreSQL driver

# Alpaca Trading
alpaca-py>=0.21.0

# Polygon/Massive Data
massive>=2.0.0
httpx>=0.25.0

# Data Processing
pandas>=2.1.3
numpy>=1.26.2
duckdb>=1.1.0
pyarrow>=12.0.1

# Environment & Config
python-dotenv>=1.0.0

# CORS & WebSockets
python-multipart>=0.0.6
websockets>=12.0

# Logging & Monitoring
structlog>=23.2.0

# Scheduling
apscheduler>=3.10.0
